Adding initial working visualization for PE #001.
Project Euler Problem 1 (https://projecteuler.net/problem=1) asks the problem solver to "Find the sum of all the multiples of 3 or 5 below 1000." This visualization draws cubes representing numbers 1 to 999. Cubes are blue when their corresponding number is divisible by 3, red when their number is divisible by 5 and purple if the number is divisible by both 3 and 5. The running sum of the numbers is shown in the lower right hand corner. Once the total is found, the application ends after closes after an additional 60s.